restructured code to make clearer the demonstration flows that are present
