+++ src/ssi/InteractionExchangeVC.ts
@@ -1,7 +1,7 @@
-import { IdentityManager } from './IdentityManager';
-import typeormConfig from './agents-ormconfig';
+import { IdentityManager } from './util/IdentityManager';
+import typeormConfig from './config/agents-ormconfig';
 import { ConnectionOptions, createConnection } from 'typeorm';
-import metadata from './SimpleCredentialMetaData'
+import metadata from './credentials/SimpleCredentialMetaData'
 import { SignedCredential } from 'jolocom-lib/js/credentials/signedCredential/signedCredential';
 import fs from "fs";
 
@@ -13,18 +13,17 @@ async function init() {
     const conf = typeormConfig as ConnectionOptions;
     const typeormConnection = await createConnection(conf);
     const identityMgr = new IdentityManager(credentialsFile, typeormConnection);
+    const logger = identityMgr.logger;
 
     //Create/load first agent
-    const password = readlineSync.question('Enter your password to either load or create new DID:').toString();
-    const alice = await identityMgr.loadBasedOnPass(password);
-    console.log(alice.identityWallet.did)
+    const alice = await identityMgr.createAgent("alicePW");
+    logger.info(`Alice: ${alice.identityWallet.did}`);
 
     //Create/load another agent
-    const password2 = readlineSync.question('Enter another password create/load second DID:').toString();
-    const bob = await identityMgr.loadBasedOnPass(password2);
-    console.log(bob.identityWallet.did)
+    const bob = await identityMgr.createAgent("bobPW");
+    logger.info(`Bob: ${bob.identityWallet.did}`);
 
-    identityMgr.logger.info("About to verify metadata between our 2 loaded/created agents")
+    logger.info("About to verify metadata between our 2 loaded/created agents...")
 
     // Alice creates the offer to Bob to sign a simple credential
     const aliceCredOffer = await alice.credOfferToken({
@@ -43,17 +42,18 @@ async function init() {
       [{ type: 'SimpleExampleCredential' }],
     );
     // Note that all agents need to also process the tokens they generate so that their interaction manager has seen all messages
-    bob.processJWT(bobCredExchangeResponse.encode());
+    await bob.processJWT(bobCredExchangeResponse.encode());
 
     // Alice receives the token response from Bob, finds the interaction + then creates the VC to share
     const aliceCredExchangeInteraction = await alice.processJWT(bobCredExchangeResponse.encode())
 
     // Create the VC that then will be issued by Alice to Bob, so that Bob can then prove that Alice attested to this credential about him. 
+    const age = readlineSync.question('Enter the age of Bob: ').toString();
     const aliceAboutBobVC = await alice.signedCredential({
       metadata: metadata,
       subject: bob.identityWallet.did,
       claim: {
-        age: 25,
+        age: age,
         name: 'Bob',
       },
     });
@@ -72,11 +72,11 @@ async function init() {
     const state: any = bobCredExchangeInteraction2.flow.state;
     
     if (state.credentialsAllValid) {
-        identityMgr.logger.info("Issued credential interaction is valid!");
+      logger.info("Issued credential interaction is valid!");
         for (let i = 0; i < state.issued.length; i++) {
           const vc = state.issued[i];
           await bob.storage.store.verifiableCredential(vc);
-          identityMgr.logger.info(`Saving verfied credential: ${JSON.stringify(vc)}`);
+          logger.info(`Saving verfied credential with claim: ${JSON.stringify(vc.claim)}`);
         }
    
         await fs.appendFileSync("VerfiedCredential.json", JSON.stringify(state)+"\r\n");
