+++ src/ssi/credentials/SimpleCredentialMetaData.ts
