+++ src/ssi/util/IdentityManager.ts
